implemented Oauth
used google Oauth for 3rd party authentication

+++ graphql/resolver.js
@@ -1,9 +1,16 @@
 const bcryptjs = require("bcryptjs");
 const User = require("../models/user");
 const validator = require("validator");
+const passport = require("passport");
+const authenticate = require("../middleware/authenticate");
 
 module.exports = {
   createUser: async function ({ userInput }, req) {
+    //authentication
+    if (!req.isAuthenticated()) {
+      throw new Error("You must be authenticated to perform this action.");
+    }
+
     //validate email
     const errors = [];
     if (!validator.isEmail(userInput.email)) {
@@ -37,38 +44,52 @@ module.exports = {
   },
 
   updateUser: async function ({ userId, userInput }, req) {
+    //auth
+    if (!req.isAuthenticated()) {
+      throw new Error("You must be authenticated to perform this action.");
+    }
+
     //validate email
-    try{
-    const errors = []
-    if(!validator.isEmail(userInput.email)){
-      errors.push({message:"Email is invalid"});
+    try {
+      const errors = [];
+      if (!validator.isEmail(userInput.email)) {
+        errors.push({ message: "Email is invalid" });
+      }
+      if (
+        validator.isEmpty(userInput.password) ||
+        !validator.isLength(userInput.password, { min: 5 })
+      ) {
+        errors.push({ message: "Please provide password." });
+      }
+      if (errors.length > 0) {
+        const error = new Error("Invalid input");
+        error.data = errors;
+        throw error;
+      }
+
+      const hashedPassword = await bcryptjs.hash(userInput.password, 12);
+
+      const user = await User.findOneAndUpdate(
+        { _id: userId.id },
+        {
+          $set: {
+            email: userInput.email,
+            name: userInput.name,
+            password: hashedPassword,
+          },
+        },
+        { new: true }
+      );
+
+      if (!user) {
+        const error = new Error("User not found");
+        throw error;
+      }
+
+      return { ...user._doc, _id: user.id };
+    } catch (err) {
+      throw err;
     }
-    if(validator.isEmpty(userInput.password) || !validator.isLength(userInput.password,{min:5})){
-      errors.push({message:"Please provide password."})
-    }
-    if (errors.length > 0){
-      const error = new Error('Invalid input')
-      error.data = errors;
-      throw error
-    }
-   
-    const hashedPassword = await bcryptjs.hash(userInput.password, 12);
-    
-    const user = await User.findOneAndUpdate(
-      { _id: userId.id },
-      { $set: { email: userInput.email, name: userInput.name, password: hashedPassword } },
-      { new: true }
-    );
-    
-    if (!user) {
-      const error = new Error('User not found');
-      throw error;
-    }
-  
-    return { ...user._doc, _id: user.id };
-  }catch(err){
-    throw err
-  }
   },
 
   getAllUsers: async function () {
@@ -88,6 +109,10 @@ module.exports = {
   },
 
   deleteUser: async function ({ userId }, req) {
+    //auth
+    if (!req.isAuthenticated()) {
+      throw new Error("You must be authenticated to perform this action.");
+    }
     try {
       await User.deleteOne({ _id: userId.id });
       return "user deleted";
